requirements.txt
runpod==1.3.0
transformers>=4.35.0
pillow>=10.0.0
numpy>=1.24.0
einops>=0.7.0
hf_transfer
boto3>=1.28.0